requirements_no_versions.txt
absl-py
aio-pika
aiofiles
aiogram
aiohappyeyeballs
aiohttp
aiohttp-retry
aiormq
aiosignal
amqp
annotated-types
anyio
appdirs
APScheduler
arabic-reshaper
arrow
asgiref
asn1crypto
astunparse
async-timeout
attrs
aubio
autobahn
Automat
Babel
bcrypt
beautifulsoup4
bidict
billiard
black
bleach
blinker
boto3
botocore
bs4
CacheControl
cachelib
cachetools
canvasapi
celery
certifi
cffi
channels
channels_redis
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudpickle
colorama
colorclass
coloredlogs
colorhash
command-not-found
confluent-kafka
constantly
contourpy
cors
cryptography
cssselect2
cycler
daphne
dask
dbus-python
decorator
deepseek-ai
Deprecated
deprecation
distlib
distro
distro-info
Django
django-auto-logout
django-cors-headers
django-extensions
django-filter
django-jazzmin
django-prometheus
django-widget-tweaks
djangorestframework
djangorestframework_simplejwt
dnspython
docopt
essentia
et_xmlfile
exceptiongroup
fabric
Fabric3
fbmessenger
ffmpeg-python
filelock
fire
flake8
flasgger
Flask
Flask-Babel
Flask-Cors
Flask-Session
flatbuffers
fonttools
frozenlist
fsspec
future
gast
gevent
google-auth
google-auth-oauthlib
google-pasta
gotrue
greenlet
grpcio
h11
h2
h5py
hpack
html2text
html5lib
httpcore
httplib2
httptools
httpx
humanfriendly
hyperframe
hyperlink
idna
importlib-metadata
incremental
iniconfig
invoke
ipython_genutils
itsdangerous
jax
jaxlib
jeepney
Jinja2
jiter
jmespath
joblib
jsonify
jsonpickle
jsonschema
jsonschema-specifications
jupyter-core
keras
keyring
kiwisolver
kombu
launchpadlib
lazr.restfulclient
lazr.uri
libclang
locket
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mattermostwrapper
mccabe
mdurl
mistune
ml-dtypes
more-itertools
mpmath
msgpack
msgspec
multidict
mypy
mypy_extensions
mysql-connector
mysql-connector-python
mysqlclient
namex
netifaces
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
openai
opencv-python
openpyxl
opt-einsum
optree
oscrypto
packaging
pamqp
pandas
paramiko
partd
pathspec
pdfminer.six
pdfplumber
pgvector
pillow
pinecone
pinecone-plugin-assistant
pinecone-plugin-interface
platformdirs
pluggy
portalocker
postgrest
prettytable
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
psycopg2-binary
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pydot
pydub
pyflakes
pygame
Pygments
PyGObject
pyHanko
pyhanko-certvalidator
PyJWT
pykwalify
pymongo
PyMuPDF
PyMySQL
PyNaCl
pyngrok
pyOpenSSL
pyotp
pyparsing
pypdf
PyPDF2
pypdfium2
PyQt5
PyQt5-Qt5
PyQt5-sip
pyrsistent
PySocks
pytest
pytest-django
python-apt
python-bidi
python-crfsuite
python-dateutil
python-decouple
python-docx
python-dotenv
python-engineio
python-magic
python-socketio
python-vlc
pytz
PyYAML
qrcode
questionary
randomname
rasa-sdk
realtime
redis
referencing
regex
reportlab
requests
requespsycopg2-binaryts-file
requests-oauthlib
requests-toolbelt
rest-framework-simplejwt
rich
rocketchat-API
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
sanic
Sanic-Cors
sanic-jwt
sanic-routing
scikit-learn
scipy
SecretStorage
sentry-sdk
service-identity
simple-websocket
simpleaudio
six
sklearn-crfsuite
slack_sdk
sniffio
soupsieve
SQLAlchemy
sqlparse
storage3
StrEnum
structlog
structlog-sentry
supabase
supafunc
svglib
sympy
systemd-python
tabulate
tarsafe
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-hub
tensorflow-io-gcs-filesystem
tensorflow-text
termcolor
terminaltables
threadpoolctl
tinycss2
tldextract
tomli
toolz
torch
tqdm
traitlets
triton
twilio
Twisted
txaio
typed-ast
types-aiofiles
types-annoy
types-appdirs
types-atomicwrites
types-aws-xray-sdk
types-babel
types-backports-abc
types-backports.ssl-match-hostname
types-beautifulsoup4
types-bleach
types-boto
types-braintree
types-cachetools
types-caldav
types-certifi
types-characteristic
types-chardet
types-click
types-click-spinner
types-colorama
types-commonmark
types-contextvars
types-croniter
types-cryptography
types-dataclasses
types-dateparser
types-DateTimeRange
types-decorator
types-Deprecated
types-docopt
types-docutils
types-editdistance
types-emoji
types-entrypoints
types-enum34
types-filelock
types-first
types-Flask
types-freezegun
types-frozendict
types-futures
types-html5lib
types-httplib2
types-humanfriendly
types-ipaddress
types-itsdangerous
types-JACK-Client
types-Jinja2
types-jmespath
types-jsonschema
types-Markdown
types-MarkupSafe
types-mock
types-mypy-extensions
types-mysqlclient
types-oauthlib
types-orjson
types-paramiko
types-Pillow
types-polib
types-prettytable
types-protobuf
types-psutil
types-psycopg2
types-pyaudio
types-pycurl
types-pyfarmhash
types-Pygments
types-PyMySQL
types-pyOpenSSL
types-pyRFC3339
types-pysftp
types-pytest-lazy-fixture
types-python-dateutil
types-python-gflags
types-python-nmap
types-python-slugify
types-pytz
types-pyvmomi
types-PyYAML
types-redis
types-requests
types-retry
types-selenium
types-Send2Trash
types-setuptools
types-simplejson
types-singledispatch
types-six
types-slumber
types-stripe
types-tabulate
types-termcolor
types-toml
types-toposort
types-ttkthemes
types-typed-ast
types-tzlocal
types-ujson
types-vobject
types-waitress
types-Werkzeug
types-xxhash
typing-inspection
typing-utils
typing_extensions
tzdata
tzlocal
ubuntu-advantage-tools
ufw
ujson
unattended-upgrades
uritools
urllib3
urwid
uuid
uvloop
vine
virtualenv
wadllib
wcwidth
webencodings
webexteamssdk
websockets
Werkzeug
wrapt
wsproto
xhtml2pdf
XlsxWriter
xlwt
yarl
zipp
zope.event
zope.interface
